moved component to it's own folder
